netlify/functions/search/requirements.txt
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
rapidfuzz==3.5.2
neo4j==5.14.0
python-dotenv==1.0.0
